requirements.txt
numpy~=1.20.1
scikit-image~=0.16.2
pillow~=7.2.0
torchvision~=0.8.1+cu101

torch~=1.7.0+cu101
argparse~=1.4.0
ipdb~=0.13.4
tqdm~=4.43.0